requirements.txt
openai
python-dotenv
pyttsx3
speechrecognition
pyaudio
pydub
requests
flask
twilio
